Removed old tests and added base code for ned versions
