--- selenium/charTest.py
@@ -1,44 +1,0 @@
-import unittest
-import time
-from selenium import webdriver
-
-class Testing(unittest.TestCase):
-	def setUp(self):
-		self.browser = webdriver.Chrome()
-		self.addCleanup(self.browser.quit)
-
-	def testCharForm(self):
-		driver = self.browser
-		driver.get('http://localhost:8000/splashHome.html')
-		time.sleep(10)
-		driver.find_element_by_id('button1').click()
-		time.sleep(4)
-		driver.find_element_by_id('charOpen').click()
-		time.sleep(2)
-		driver.find_element_by_id('formButtonChar').click()
-		driver.find_element_by_id('characterName').send_keys('Haladar')
-		driver.find_element_by_id('HP').send_keys('65')
-		driver.find_element_by_id('AC').send_keys('15')
-		driver.find_element_by_id('PP').send_keys('15')
-		driver.find_element_by_id('Strength').send_keys('8')
-		driver.find_element_by_id('Dexterity').send_keys('2')
-		driver.find_element_by_id('Constitution').send_keys('6')
-		driver.find_element_by_id('Intelligence').send_keys('1')
-		driver.find_element_by_id('Wisdom').send_keys('1')
-		driver.find_element_by_id('Charisma').send_keys('0')
-		driver.find_element_by_id('PlayerName').send_keys('Max')
-		driver.find_element_by_id('Background').send_keys('Made of cheese')
-		driver.find_element_by_id('ClassAndLevel').send_keys('Cleric 10')
-		driver.find_element_by_id('Experience').send_keys('55')
-		driver.find_element_by_id('Alignment').send_keys('Chaotic Neutral')
-		driver.find_element_by_id('Race').send_keys('Martian')
-		time.sleep(2)
-		driver.find_element_by_id('doneChar').click()
-		time.sleep(3)
-
-	def tearDown(self):
-		self.browser.close()
-
-
-if __name__ == '__main__':
-    unittest.main(verbosity=2)
